Update codecs. Correct data types.

+++ common/src/main/scala/io/apex/f1/data/FastestLap.scala
@@ -1,5 +1,5 @@
 package io.apex.f1.data
 
-case class FastestLap(vehicleIdx: Short, lapTime: Float) {
+case class FastestLap(vehicleIdx: Int, lapTime: Float) {
   override def toString: String = s"FastestLap[vehicleIdx=$vehicleIdx, lapTime=$lapTime]"
 }
